* Compiler warning
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@1244 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ player.cpp
@@ -3814,7 +3814,7 @@ void Player::setPromotionLevel(uint32_t pLevel)
 double Player::getLostPercent(lossTypes_t lossType)
 {
 	uint32_t lostPercent = lossPercent[lossType];
-	if(lostPercent <= vocation->getLessLoss())
+	if((int32_t)lostPercent <= vocation->getLessLoss())
 		return 0;
 
 	lostPercent -= vocation->getLessLoss();
